requirements.txt
accelerate==1.12.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
boto3==1.42.13
botocore==1.42.13
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
cryptography==46.0.3
debugpy==1.8.19
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
exceptiongroup==1.3.1
executing==2.2.1
faiss-cpu==1.7.4
filelock==3.20.1
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
groq==1.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.21.4
idna==3.11
importlib-metadata==6.11.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
openai==2.13.0
packaging==23.2
pandas==2.3.3
parso==0.8.5
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyPDF2==3.0.1
pypdfium2==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.4.2
scikit-learn==1.7.2
scipy==1.11.4
sentence-transformers==2.3.1
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.29.0
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.15.2
toml==0.10.2
torch==2.1.2
torchvision==0.16.2
tornado==6.5.4
tqdm==4.66.1
traitlets==5.14.3
transformers==4.37.2
triton==2.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.14
zipp==3.23.0
